Move ButlerURI to separate file
