+++ tests/test_location.py
@@ -24,7 +24,7 @@ import os.path
 import posixpath
 
 from lsst.daf.butler import LocationFactory, ButlerURI
-from lsst.daf.butler.core.location import os2posix, posix2os
+from lsst.daf.butler.core._butlerUri import os2posix, posix2os
 
 
 class LocationTestCase(unittest.TestCase):
